* commit '3a4e905a92c6fdc14d116be2b0dc7352ac3b1cc2':
  GYRE-1236 fix code style
  GYRE-1236 исправления по комментам МР
  GYRE-1236 нотификации сброс пароля и верификация емейл
  GYRE-1236 остальные письма из Notification.php
  GYRE-1236 тест письма "мало комметов в этом месяце, пароль через социалки, давно не посещали сайт(месячное письмо)"
  GYRE-1236 тест письма "мало комментариев на этой неделе"
  GYRE-1147 фабрики для тестов
  GYRE-1236 Подготовка и отладка тестов

+++ database/migrations/2016_01_04_173148_create_admin_tables.php
@@ -95,8 +95,13 @@ class CreateAdminTables extends Migration
 
         Schema::connection($connection)->table(config('admin.database.users_table'), function (Blueprint $table) {
             $table->renameColumn('username', 'name');
-            $table->dropColumn('avatar');
         });
+
+        if (Schema::hasColumn(config('admin.database.users_table'), 'avatar')) {
+            Schema::connection($connection)->table(config('admin.database.users_table'), function (Blueprint $table) {
+                $table->dropColumn('avatar');
+            });
+        }
 
         Schema::connection($connection)->table(config('admin.database.users_table'), function (Blueprint $table) {
             $table->string('name', 190)->change();
